// src/pages/InvoiceManage/typings.d.ts
// @ts-ignore
/* eslint-disable */

declare namespace API {
  type Params = {
    customerName: string,
    pageNumber: number,
    pageSize: number,
    status: number
  };

  type TableItem = {
    customerName: string,
    customerNameAddress: string,
    invoiceAmount: number,
    invoiceDate: string,
    invoiceRecordCode: string,
    invoiceType: number,
    operatorName: string,
    status: number
  }
}